pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "devdiary"
version = "0.1.0"
description = "DevDiary - Your Second Brain for Development. Automatically generate dev diaries from Git history with AI."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "DevDiary Team"},
]
keywords = ["git", "diary", "ai", "developer-tools", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "aiosqlite>=0.19.0",
    "alembic>=1.13.0",
    "gitpython>=3.1.40",
    "click>=8.1.7",
    "rich>=13.7.0",
    "pyyaml>=6.0.1",
    "httpx>=0.26.0",
    "jinja2>=3.1.3",
    "python-multipart>=0.0.6",
    "markdown>=3.5.2",
    "pygments>=2.17.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "httpx>=0.26.0",
]
postgres = [
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
]

[project.scripts]
devdiary = "devdiary.cli.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.flake8]
max-line-length = 100
exclude = [".git", "__pycache__", "build", "dist", "alembic"]
